Added generating coverage report
