+++ .circleci/config.yml
@@ -3,6 +3,8 @@ version: 2
 jobs:
   test:
     working_directory: ~/check-flow
+    environment:
+      CC_TEST_REPORTER_ID: YOUR_REPO_CC_TEST_REPORTER_ID
     docker:
       - image: circleci/node:8
     steps:
@@ -11,8 +13,17 @@ jobs:
           name: Install Packages
           command: yarn install
       - run:
+          name: Setup Code Climate test-reporter
+          command: |
+            curl -L https://codeclimate.com/downloads/test-reporter/test-reporter-latest-linux-amd64 > ./cc-test-reporter
+            chmod +x ./cc-test-reporter
+            ./cc-test-reporter before-build
+      - run:
           name: Test
-          command: yarn run test
+          command: yarn run test:coverage
+      - run:
+          name: Report Coverage
+          command: ./cc-test-reporter after-build --coverage-input-type lcov --exit-code $?
 
   lint:
     working_directory: ~/check-flow
