+++ .idea/dictionaries/henri.xml
@@ -1,0 +1,3 @@
+<component name="ProjectDictionaryState">
+  <dictionary name="henri" />
+</component>
