+++ README.md
@@ -1,6 +1,6 @@
 # check-flow
 
-[![npm](https://img.shields.io/npm/v/check-flow.svg)](https://www.npmjs.com/package/check-flow) ![Maintenance](https://img.shields.io/maintenance/yes/2018.svg) ![npm](https://img.shields.io/npm/l/check-flow.svg) ![David](https://img.shields.io/david/HenriBeck/check-flow.svg) [![CircleCI](https://circleci.com/gh/HenriBeck/check-flow.svg?style=svg)](https://circleci.com/gh/HenriBeck/check-flow)
+[![npm](https://img.shields.io/npm/v/check-flow.svg)](https://www.npmjs.com/package/check-flow) ![Maintenance](https://img.shields.io/maintenance/yes/2018.svg) ![npm](https://img.shields.io/npm/l/check-flow.svg) ![David](https://img.shields.io/david/HenriBeck/check-flow.svg) [![CircleCI](https://circleci.com/gh/HenriBeck/check-flow.svg?style=svg)](https://circleci.com/gh/HenriBeck/check-flow) [![Maintainability](https://api.codeclimate.com/v1/badges/0ddecaa2ad0f3faf830e/maintainability)](https://codeclimate.com/github/HenriBeck/check-flow/maintainability) [![Test Coverage](https://api.codeclimate.com/v1/badges/0ddecaa2ad0f3faf830e/test_coverage)](https://codeclimate.com/github/HenriBeck/check-flow/test_coverage)
 
 A better flow cli with ignoring errors from files which are in `node_modules` folder.
